requirements.txt
firecrawl-py
google-genai
requests
pandas
streamlit
plotly
python-dotenv
sqlite-utils
